+++ tech1-framework-emails/src/main/java/io/tech1/framework/emails/domain/EmailPlainAttachment.java
@@ -2,13 +2,13 @@ package io.tech1.framework.emails.domain;
 
 import lombok.Data;
 
-import java.util.Map;
 import java.util.Set;
 
 @Data
-public class EmailHTML {
+public class EmailPlainAttachment {
     private final Set<String> to;
     private final String subject;
-    private final String templateName;
-    private final Map<String, Object> templateVariables;
+    private final String message;
+    private final String attachmentFileName;
+    private final String attachmentMessage;
 }