+++ tech1-framework-emails/src/main/java/io/tech1/framework/emails/services/EmailService.java
@@ -1,6 +1,7 @@
 package io.tech1.framework.emails.services;
 
 import io.tech1.framework.emails.domain.EmailHTML;
+import io.tech1.framework.emails.domain.EmailPlainAttachment;
 import org.springframework.scheduling.annotation.Async;
 
 import java.util.List;
@@ -15,6 +16,9 @@ public interface EmailService {
     void sendPlain(List<String> to, String subject, String message);
     @Async
     void sendPlain(Set<String> to, String subject, String message);
+
+    @Async
+    void sendPlainAttachment(EmailPlainAttachment emailPlainAttachment);
 
     @Async
     void sendHTML(EmailHTML emailHTML);